+++ beacon-chain/state/state-native/getters_gloas.go
@@ -6,7 +6,6 @@ import (
 	fieldparams "github.com/OffchainLabs/prysm/v7/config/fieldparams"
 	"github.com/OffchainLabs/prysm/v7/config/params"
 	"github.com/OffchainLabs/prysm/v7/consensus-types/primitives"
-	enginev1 "github.com/OffchainLabs/prysm/v7/proto/engine/v1"
 	ethpb "github.com/OffchainLabs/prysm/v7/proto/prysm/v1alpha1"
 	"github.com/OffchainLabs/prysm/v7/runtime/version"
 )
@@ -47,6 +46,14 @@ func (b *BeaconState) BuilderPubkey(builderIndex primitives.BuilderIndex) ([fiel
 }
 
 // IsActiveBuilder returns true if the builder placement is finalized and it has not initiated exit.
+// Spec v1.7.0-alpha.0 (pseudocode):
+// def is_active_builder(state: BeaconState, builder_index: BuilderIndex) -> bool:
+//
+//	builder = state.builders[builder_index]
+//	return (
+//	    builder.deposit_epoch < state.finalized_checkpoint.epoch
+//	    and builder.withdrawable_epoch == FAR_FUTURE_EPOCH
+//	)
 func (b *BeaconState) IsActiveBuilder(builderIndex primitives.BuilderIndex) (bool, error) {
 	if b.version < version.Gloas {
 		return false, errNotSupported("IsActiveBuilder", b.version)
@@ -65,6 +72,15 @@ func (b *BeaconState) IsActiveBuilder(builderIndex primitives.BuilderIndex) (boo
 }
 
 // CanBuilderCoverBid returns true if the builder has enough balance to cover the given bid amount.
+// Spec v1.7.0-alpha.0 (pseudocode):
+// def can_builder_cover_bid(state: BeaconState, builder_index: BuilderIndex, bid_amount: Gwei) -> bool:
+//
+//	builder_balance = state.builders[builder_index].balance
+//	pending_withdrawals_amount = get_pending_balance_to_withdraw_for_builder(state, builder_index)
+//	min_balance = MIN_DEPOSIT_AMOUNT + pending_withdrawals_amount
+//	if builder_balance < min_balance:
+//	    return False
+//	return builder_balance - min_balance >= bid_amount
 func (b *BeaconState) CanBuilderCoverBid(builderIndex primitives.BuilderIndex, bidAmount primitives.Gwei) (bool, error) {
 	if b.version < version.Gloas {
 		return false, errNotSupported("CanBuilderCoverBid", b.version)
@@ -89,19 +105,6 @@ func (b *BeaconState) CanBuilderCoverBid(builderIndex primitives.BuilderIndex, b
 	return balance-minBalance >= uint64(bidAmount), nil
 }
 
-// executionPayloadAvailabilityVal returns a copy of the execution payload availability.
-// This assumes that a lock is already held on BeaconState.
-func (b *BeaconState) executionPayloadAvailabilityVal() []byte {
-	if b.executionPayloadAvailability == nil {
-		return nil
-	}
-
-	availability := make([]byte, len(b.executionPayloadAvailability))
-	copy(availability, b.executionPayloadAvailability)
-
-	return availability
-}
-
 // BuilderPendingPaymentsNoCopy returns the builder pending payments without copying.
 func (b *BeaconState) BuilderPendingPaymentsNoCopy() ([]*ethpb.BuilderPendingPayment, error) {
 	if b.version < version.Gloas {
@@ -112,80 +115,6 @@ func (b *BeaconState) BuilderPendingPaymentsNoCopy() ([]*ethpb.BuilderPendingPay
 	defer b.lock.RUnlock()
 
 	return b.builderPendingPayments, nil
-}
-
-// builderPendingPaymentsVal returns a copy of the builder pending payments.
-// This assumes that a lock is already held on BeaconState.
-func (b *BeaconState) builderPendingPaymentsVal() []*ethpb.BuilderPendingPayment {
-	if b.builderPendingPayments == nil {
-		return nil
-	}
-
-	payments := make([]*ethpb.BuilderPendingPayment, len(b.builderPendingPayments))
-	for i, payment := range b.builderPendingPayments {
-		payments[i] = payment.Copy()
-	}
-
-	return payments
-}
-
-// builderPendingWithdrawalsVal returns a copy of the builder pending withdrawals.
-// This assumes that a lock is already held on BeaconState.
-func (b *BeaconState) builderPendingWithdrawalsVal() []*ethpb.BuilderPendingWithdrawal {
-	if b.builderPendingWithdrawals == nil {
-		return nil
-	}
-
-	withdrawals := make([]*ethpb.BuilderPendingWithdrawal, len(b.builderPendingWithdrawals))
-	for i, withdrawal := range b.builderPendingWithdrawals {
-		withdrawals[i] = withdrawal.Copy()
-	}
-
-	return withdrawals
-}
-
-// buildersVal returns a copy of the builders registry.
-// This assumes that a lock is already held on BeaconState.
-func (b *BeaconState) buildersVal() []*ethpb.Builder {
-	if b.builders == nil {
-		return nil
-	}
-
-	builders := make([]*ethpb.Builder, len(b.builders))
-	for i := range builders {
-		builder := b.builders[i]
-		builders[i] = ethpb.CopyBuilder(builder)
-	}
-
-	return builders
-}
-
-// latestBlockHashVal returns a copy of the latest block hash.
-// This assumes that a lock is already held on BeaconState.
-func (b *BeaconState) latestBlockHashVal() []byte {
-	if b.latestBlockHash == nil {
-		return nil
-	}
-
-	hash := make([]byte, len(b.latestBlockHash))
-	copy(hash, b.latestBlockHash)
-
-	return hash
-}
-
-// payloadExpectedWithdrawalsVal returns a copy of the payload expected withdrawals.
-// This assumes that a lock is already held on BeaconState.
-func (b *BeaconState) payloadExpectedWithdrawalsVal() []*enginev1.Withdrawal {
-	if b.payloadExpectedWithdrawals == nil {
-		return nil
-	}
-
-	withdrawals := make([]*enginev1.Withdrawal, len(b.payloadExpectedWithdrawals))
-	for i, withdrawal := range b.payloadExpectedWithdrawals {
-		withdrawals[i] = withdrawal.Copy()
-	}
-
-	return withdrawals
 }
 
 // builderAtIndex intentionally returns the underlying pointer without copying.
